Rake now supports a --jobs <n> command-line option.
--jobs number (-j)
Specifies the maximum number of concurrent tasks. The suggested value is equal to the number of CPUs.

Sample values:
  default: unlimited concurrent tasks (standard rake behavior)
  1: one task at a time

The code consists of two major edits, the first is a change to
`application.rb` to support the parsing of the option.

The second is a more substantial change to `multi_task.rb` which
replaces the multi-task scheduling algorithm. Instead of spawning
a new thread for every pre-requisite that needs to be executed,
a Proc is created which calls the pre-requisite and added to a Queue.

Additionally, a thread-pool is created to pull the Procs off the queue
and execute them. Finally, the MultiTask queueing up its prerequisites
will participate in the Proc-processing while it waits for its
prerequisites to finish processing.

It can tell when its prerequisites are finished by adding a little
bookkeeping to the queued-up Proc objects.